Try building a patched Ruby.

+++ build.sh
@@ -61,8 +61,13 @@ fold_end rvm.2
 fold_start build "build $RUBY"
 rvm alias delete $RUBY
 rvm remove $RUBY
-rvm install $RUBY --verify-downloads 1
-rvm prepare $RUBY
+if [ -e "${RUBY_PATCH}" ]; then
+  rvm install $RUBY --patch $RUBY_PATCH -n $RUBY_PATCH --verify-downloads 1
+  rvm prepare $RUBY
+else
+  rvm install $RUBY --verify-downloads 1
+  rvm prepare $RUBY
+fi
 fold_end build
 
 #######################################################
